refactor: split schema/schemahandler.go into handler and builder files
Extract schema-building logic into schemabuild.go; split matching tests
into schemabuild_test.go; move gettype tests to gettype_test.go.
