+++ schema/schemahandler_test.go
@@ -9,363 +9,6 @@ import (
 	"github.com/hangxie/parquet-go/v3/parquet"
 )
 
-// Tests for schema handler functionality
-
-func TestNewSchemaHandlerFromSchemaHandler(t *testing.T) {
-	// Create an original schema using NewSchemaHandlerFromStruct for simplicity
-	originalSchema, err := NewSchemaHandlerFromStruct(new(struct {
-		Field1 int32  `parquet:"name=field1, type=INT32"`
-		Field2 string `parquet:"name=field2, type=BYTE_ARRAY, convertedtype=UTF8, repetitiontype=OPTIONAL"`
-	}))
-	require.NoError(t, err)
-
-	// Test NewSchemaHandlerFromSchemaHandler
-	schema2 := NewSchemaHandlerFromSchemaHandler(originalSchema)
-
-	// Verify that the schema was copied correctly
-	require.Equal(t, len(originalSchema.SchemaElements), len(schema2.SchemaElements))
-	require.Equal(t, common.ParGoRootInName, schema2.SchemaElements[0].Name)
-	require.Equal(t, "Field1", schema2.SchemaElements[1].Name)
-	require.Equal(t, "Field2", schema2.SchemaElements[2].Name)
-
-	// Verify the schema elements were deep copied
-	require.Equal(t, "REQUIRED", schema2.SchemaElements[1].RepetitionType.String())
-	require.Equal(t, "INT32", schema2.SchemaElements[1].Type.String())
-
-	require.Equal(t, "OPTIONAL", schema2.SchemaElements[2].RepetitionType.String())
-	require.Equal(t, "BYTE_ARRAY", schema2.SchemaElements[2].Type.String())
-	require.Equal(t, "UTF8", schema2.SchemaElements[2].ConvertedType.String())
-
-	// Verify it's a separate copy, not the same reference
-	require.NotSame(t, originalSchema, schema2)
-	require.NotSame(t, &originalSchema.SchemaElements, &schema2.SchemaElements)
-}
-
-func TestNewSchemaHandlerFromSchemaList(t *testing.T) {
-	// Create a simple schema element list directly
-	requiredRep := parquet.FieldRepetitionType_REQUIRED
-	optionalRep := parquet.FieldRepetitionType_OPTIONAL
-	int32Type := parquet.Type_INT32
-	byteArrayType := parquet.Type_BYTE_ARRAY
-	utf8Conv := parquet.ConvertedType_UTF8
-	numChildren := int32(2)
-
-	schemaElements := []*parquet.SchemaElement{
-		{Name: common.ParGoRootInName, RepetitionType: &requiredRep, NumChildren: &numChildren},
-		{Name: "Field1", Type: &int32Type, RepetitionType: &requiredRep},
-		{Name: "Field2", Type: &byteArrayType, ConvertedType: &utf8Conv, RepetitionType: &optionalRep},
-	}
-
-	schema2 := NewSchemaHandlerFromSchemaList(schemaElements)
-
-	// Verify that the schema was created correctly from the list
-	require.Equal(t, len(schemaElements), len(schema2.SchemaElements))
-	require.Equal(t, common.ParGoRootInName, schema2.SchemaElements[0].Name)
-	require.Equal(t, "Field1", schema2.SchemaElements[1].Name)
-	require.Equal(t, "Field2", schema2.SchemaElements[2].Name)
-
-	// Verify the schema elements properties
-	require.Equal(t, "REQUIRED", schema2.SchemaElements[1].RepetitionType.String())
-	require.Equal(t, "INT32", schema2.SchemaElements[1].Type.String())
-
-	require.Equal(t, "OPTIONAL", schema2.SchemaElements[2].RepetitionType.String())
-	require.Equal(t, "BYTE_ARRAY", schema2.SchemaElements[2].Type.String())
-	require.Equal(t, "UTF8", schema2.SchemaElements[2].ConvertedType.String())
-
-	// Verify that maps were created properly
-	require.NotNil(t, schema2.MapIndex)
-	require.NotNil(t, schema2.PathMap)
-}
-
-func TestNewSchemaHandlerFromStruct(t *testing.T) {
-	tests := []struct {
-		name              string
-		structDef         interface{}
-		expectedErrorText string
-		validateSchema    func(t *testing.T, schema *SchemaHandler)
-	}{
-		{
-			name: "invalid_tag",
-			structDef: new(struct {
-				Id int32 `parquet:"foo=bar, type=INT32"`
-			}),
-			expectedErrorText: "unrecognized tag 'foo'",
-		},
-		{
-			name: "invalid_type",
-			structDef: new(struct {
-				Name string `parquet:"name=name, type=UTF8"`
-			}),
-			expectedErrorText: "field [Name] with type [UTF8]: not a valid Type strin",
-		},
-		{
-			name: "ignore_field_without_tag",
-			structDef: new(struct {
-				Id   int32 `parquet:"name=id, type=INT32, ConvertedType=INT_32"`
-				Name string
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				require.Equal(t, 2, len(schema.SchemaElements))
-				require.Equal(t, common.ParGoRootInName, schema.SchemaElements[0].Name)
-				require.Equal(t, "Id", schema.SchemaElements[1].Name)
-				require.Equal(t, "INT32", schema.SchemaElements[1].Type.String())
-				require.Equal(t, "REQUIRED", schema.SchemaElements[1].RepetitionType.String())
-				require.True(t, schema.SchemaElements[1].LogicalType.IsSetINTEGER())
-				require.Equal(t, "INT_32", schema.SchemaElements[1].ConvertedType.String())
-			},
-		},
-		{
-			name: "pointer_field",
-			structDef: new(struct {
-				Name *string `parquet:"name=name, type=BYTE_ARRAY, ConvertedType=UTF8"`
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				require.Equal(t, 2, len(schema.SchemaElements))
-				require.Equal(t, common.ParGoRootInName, schema.SchemaElements[0].Name)
-				require.Equal(t, "Name", schema.SchemaElements[1].Name)
-				require.Equal(t, "BYTE_ARRAY", schema.SchemaElements[1].Type.String())
-				require.Equal(t, "OPTIONAL", schema.SchemaElements[1].RepetitionType.String())
-				require.True(t, schema.SchemaElements[1].LogicalType.IsSetSTRING())
-				require.Equal(t, "UTF8", schema.SchemaElements[1].ConvertedType.String())
-			},
-		},
-		{
-			name: "map_missing_value_type",
-			structDef: new(struct {
-				MapField map[string]*int32 `parquet:"name=map, type=MAP, convertedtype=MAP, keytype=BYTE_ARRAY, keyconvertedtype=UTF8"`
-			}),
-			expectedErrorText: "field [Value] with type []: not a valid Type string",
-		},
-		{
-			name: "map_missing_key_type",
-			structDef: new(struct {
-				MapField map[string]*int32 `parquet:"name=map, type=MAP, convertedtype=MAP, valuetype=INT32"`
-			}),
-			expectedErrorText: "field [Key] with type []: not a valid Type string",
-		},
-		{
-			name: "map_good",
-			structDef: new(struct {
-				MapField1 map[string]*int32  `parquet:"name=map, type=MAP, convertedtype=MAP, keytype=BYTE_ARRAY, keyconvertedtype=UTF8, valuetype=INT32"`
-				MapField2 *map[*string]int32 `parquet:"name=map, type=MAP, convertedtype=MAP, keytype=BYTE_ARRAY, keyconvertedtype=UTF8, valuetype=INT32"`
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				require.Equal(t, 9, len(schema.SchemaElements))
-				require.Equal(t, common.ParGoRootInName, schema.SchemaElements[0].Name)
-				require.Equal(t, "MapField1", schema.SchemaElements[1].Name)
-				require.Equal(t, "Key_value", schema.SchemaElements[2].Name)
-				require.Equal(t, "Key", schema.SchemaElements[3].Name)
-				require.Equal(t, "Value", schema.SchemaElements[4].Name)
-				require.Equal(t, "MapField2", schema.SchemaElements[5].Name)
-				require.Equal(t, "Key_value", schema.SchemaElements[6].Name)
-				require.Equal(t, "Key", schema.SchemaElements[7].Name)
-				require.Equal(t, "Value", schema.SchemaElements[8].Name)
-
-				require.Equal(t, "REQUIRED", schema.SchemaElements[1].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[1].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[1].Type)
-				require.Equal(t, "MAP", schema.SchemaElements[1].ConvertedType.String())
-
-				require.Equal(t, "REPEATED", schema.SchemaElements[2].RepetitionType.String())
-				require.Equal(t, int32(2), schema.SchemaElements[2].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[2].Type)
-				require.Equal(t, "MAP_KEY_VALUE", schema.SchemaElements[2].ConvertedType.String())
-
-				require.Equal(t, "REQUIRED", schema.SchemaElements[3].RepetitionType.String())
-				require.Equal(t, "BYTE_ARRAY", schema.SchemaElements[3].Type.String())
-				require.Equal(t, "UTF8", schema.SchemaElements[3].ConvertedType.String())
-
-				require.Equal(t, "OPTIONAL", schema.SchemaElements[4].RepetitionType.String())
-				require.Equal(t, "INT32", schema.SchemaElements[4].Type.String())
-				require.Nil(t, schema.SchemaElements[4].ConvertedType)
-
-				require.Equal(t, "OPTIONAL", schema.SchemaElements[5].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[5].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[5].Type)
-				require.Equal(t, "MAP", schema.SchemaElements[5].ConvertedType.String())
-
-				require.Equal(t, "REPEATED", schema.SchemaElements[6].RepetitionType.String())
-				require.Equal(t, int32(2), schema.SchemaElements[6].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[6].Type)
-				require.Equal(t, "MAP_KEY_VALUE", schema.SchemaElements[6].ConvertedType.String())
-
-				require.Equal(t, "REQUIRED", schema.SchemaElements[7].RepetitionType.String())
-				require.Equal(t, "BYTE_ARRAY", schema.SchemaElements[7].Type.String())
-				require.Equal(t, "UTF8", schema.SchemaElements[7].ConvertedType.String())
-
-				require.Equal(t, "REQUIRED", schema.SchemaElements[8].RepetitionType.String())
-				require.Equal(t, "INT32", schema.SchemaElements[8].Type.String())
-				require.Nil(t, schema.SchemaElements[8].ConvertedType)
-			},
-		},
-		{
-			name: "list_missing_element_type",
-			structDef: new(struct {
-				ListField *[]string `parquet:"name=list, type=LIST, convertedtype=LIST"`
-			}),
-			expectedErrorText: "field [Element] with type []: not a valid Type string",
-		},
-		{
-			name: "list_good",
-			structDef: new(struct {
-				ListField1 *[]string `parquet:"name=list, type=LIST, convertedtype=LIST, valuetype=BYTE_ARRAY, valueconvertedtype=UTF8"`
-				ListField2 []*string `parquet:"name=list, type=LIST, convertedtype=LIST, valuetype=BYTE_ARRAY, valueconvertedtype=UTF8"`
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				require.Equal(t, 7, len(schema.SchemaElements))
-				require.Equal(t, common.ParGoRootInName, schema.SchemaElements[0].Name)
-				require.Equal(t, "ListField1", schema.SchemaElements[1].Name)
-				require.Equal(t, "List", schema.SchemaElements[2].Name)
-				require.Equal(t, "Element", schema.SchemaElements[3].Name)
-				require.Equal(t, "ListField2", schema.SchemaElements[4].Name)
-				require.Equal(t, "List", schema.SchemaElements[5].Name)
-				require.Equal(t, "Element", schema.SchemaElements[6].Name)
-
-				require.Equal(t, "OPTIONAL", schema.SchemaElements[1].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[1].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[1].Type)
-				require.Equal(t, "LIST", schema.SchemaElements[1].ConvertedType.String())
-
-				require.Equal(t, "REPEATED", schema.SchemaElements[2].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[2].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[2].Type)
-				require.Nil(t, schema.SchemaElements[2].ConvertedType)
-
-				require.Equal(t, "REQUIRED", schema.SchemaElements[3].RepetitionType.String())
-				require.Equal(t, "BYTE_ARRAY", schema.SchemaElements[3].Type.String())
-				require.Equal(t, "UTF8", schema.SchemaElements[3].ConvertedType.String())
-
-				require.Equal(t, "REQUIRED", schema.SchemaElements[4].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[4].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[4].Type)
-				require.Equal(t, "LIST", schema.SchemaElements[4].ConvertedType.String())
-
-				require.Equal(t, "REPEATED", schema.SchemaElements[5].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[5].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[5].Type)
-				require.Nil(t, schema.SchemaElements[5].ConvertedType)
-
-				require.Equal(t, "OPTIONAL", schema.SchemaElements[6].RepetitionType.String())
-				require.Equal(t, "BYTE_ARRAY", schema.SchemaElements[6].Type.String())
-				require.Equal(t, "UTF8", schema.SchemaElements[6].ConvertedType.String())
-			},
-		},
-		{
-			name: "list_repeated",
-			structDef: new(struct {
-				ListField []int32 `parquet:"name=repeated, type=INT32, repetitiontype=REPEATED"`
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				require.Equal(t, 2, len(schema.SchemaElements))
-				require.Equal(t, common.ParGoRootInName, schema.SchemaElements[0].Name)
-				require.Equal(t, "ListField", schema.SchemaElements[1].Name)
-
-				require.Equal(t, "REPEATED", schema.SchemaElements[1].RepetitionType.String())
-				require.Equal(t, "INT32", schema.SchemaElements[1].Type.String())
-				require.Nil(t, schema.SchemaElements[1].ConvertedType)
-			},
-		},
-		{
-			name: "variant_type",
-			structDef: new(struct {
-				VariantField any `parquet:"name=variant, type=VARIANT, logicaltype=VARIANT"`
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				// VARIANT should create a GROUP with 2 children: Metadata and Value
-				require.Equal(t, 4, len(schema.SchemaElements))
-				require.Equal(t, common.ParGoRootInName, schema.SchemaElements[0].Name)
-				require.Equal(t, "VariantField", schema.SchemaElements[1].Name)
-				require.Equal(t, "Metadata", schema.SchemaElements[2].Name)
-				require.Equal(t, "Value", schema.SchemaElements[3].Name)
-
-				// Verify VARIANT group structure
-				require.Equal(t, "REQUIRED", schema.SchemaElements[1].RepetitionType.String())
-				require.Equal(t, int32(2), schema.SchemaElements[1].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[1].Type) // GROUP has no type
-				require.True(t, schema.SchemaElements[1].LogicalType.IsSetVARIANT())
-
-				// Verify Metadata child
-				require.Equal(t, "REQUIRED", schema.SchemaElements[2].RepetitionType.String())
-				require.Equal(t, "BYTE_ARRAY", schema.SchemaElements[2].Type.String())
-
-				// Verify Value child
-				require.Equal(t, "REQUIRED", schema.SchemaElements[3].RepetitionType.String())
-				require.Equal(t, "BYTE_ARRAY", schema.SchemaElements[3].Type.String())
-			},
-		},
-		{
-			name: "map_fields_comprehensive",
-			structDef: new(struct {
-				MapField1 map[string]*int32  `parquet:"name=map1, type=MAP, convertedtype=MAP, keytype=BYTE_ARRAY, keyconvertedtype=UTF8, valuetype=INT32"`
-				MapField2 *map[*string]int32 `parquet:"name=map2, type=MAP, convertedtype=MAP, keytype=BYTE_ARRAY, keyconvertedtype=UTF8, valuetype=INT32"`
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				require.Equal(t, 9, len(schema.SchemaElements))
-				require.Equal(t, common.ParGoRootInName, schema.SchemaElements[0].Name)
-				require.Equal(t, "MapField1", schema.SchemaElements[1].Name)
-				require.Equal(t, "Key_value", schema.SchemaElements[2].Name)
-				require.Equal(t, "Key", schema.SchemaElements[3].Name)
-				require.Equal(t, "Value", schema.SchemaElements[4].Name)
-				require.Equal(t, "MapField2", schema.SchemaElements[5].Name)
-				require.Equal(t, "Key_value", schema.SchemaElements[6].Name)
-				require.Equal(t, "Key", schema.SchemaElements[7].Name)
-				require.Equal(t, "Value", schema.SchemaElements[8].Name)
-
-				// Verify MapField1 structure
-				require.Equal(t, "REQUIRED", schema.SchemaElements[1].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[1].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[1].Type)
-				require.Equal(t, "MAP", schema.SchemaElements[1].ConvertedType.String())
-
-				// Verify MapField2 structure (pointer to map)
-				require.Equal(t, "OPTIONAL", schema.SchemaElements[5].RepetitionType.String())
-				require.Equal(t, int32(1), schema.SchemaElements[5].GetNumChildren())
-				require.Nil(t, schema.SchemaElements[5].Type)
-				require.Equal(t, "MAP", schema.SchemaElements[5].ConvertedType.String())
-			},
-		},
-		{
-			name: "variant_with_encoding_and_compression",
-			structDef: new(struct {
-				VariantField any `parquet:"name=variant, type=VARIANT, encoding=PLAIN, compression=GZIP"`
-			}),
-			validateSchema: func(t *testing.T, schema *SchemaHandler) {
-				require.Equal(t, 4, len(schema.SchemaElements))
-
-				// Verify Metadata child has encoding and compression
-				require.Equal(t, "Metadata", schema.SchemaElements[2].Name)
-				require.Equal(t, parquet.Encoding_PLAIN, schema.Infos[2].Encoding)
-				require.NotNil(t, schema.Infos[2].CompressionType)
-				require.Equal(t, parquet.CompressionCodec_GZIP, *schema.Infos[2].CompressionType)
-
-				// Verify Value child has encoding and compression
-				require.Equal(t, "Value", schema.SchemaElements[3].Name)
-				require.Equal(t, parquet.Encoding_PLAIN, schema.Infos[3].Encoding)
-				require.NotNil(t, schema.Infos[3].CompressionType)
-				require.Equal(t, parquet.CompressionCodec_GZIP, *schema.Infos[3].CompressionType)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			schema, err := NewSchemaHandlerFromStruct(tt.structDef)
-
-			if tt.expectedErrorText != "" {
-				require.NotNil(t, err)
-				require.Contains(t, err.Error(), tt.expectedErrorText)
-				require.Nil(t, schema)
-			} else {
-				require.Nil(t, err)
-				require.NotNil(t, schema)
-				if tt.validateSchema != nil {
-					tt.validateSchema(t, schema)
-				}
-			}
-		})
-	}
-}
-
 func TestSchemaHandler_GetRepetitionLevelIndex(t *testing.T) {
 	// Create a schema with repeated and nested fields
 	schema, err := NewSchemaHandlerFromStruct(new(struct {
@@ -739,304 +382,6 @@ func TestGetRepetitionType(t *testing.T) {
 	})
 }
 
-func TestIsValidVariantSchema(t *testing.T) {
-	// Helper to create a valid VARIANT schema handler
-	createVariantSchema := func() (*SchemaHandler, int32, []int32) {
-		sh, err := NewSchemaHandlerFromStruct(new(struct {
-			VariantField any `parquet:"name=variant, type=VARIANT, logicaltype=VARIANT"`
-		}))
-		if err != nil {
-			t.Fatalf("failed to create variant schema: %v", err)
-		}
-		// Build children map
-		sh.childrenMap = sh.buildChildrenMap()
-		// VariantField is at index 1, children are Metadata (2) and Value (3)
-		return sh, 1, sh.childrenMap[1]
-	}
-
-	t.Run("valid variant schema", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		require.True(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("missing VARIANT LogicalType", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Remove the VARIANT LogicalType
-		sh.SchemaElements[idx].LogicalType = nil
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("wrong LogicalType (not VARIANT)", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Set a different LogicalType
-		sh.SchemaElements[idx].LogicalType = parquet.NewLogicalType()
-		sh.SchemaElements[idx].LogicalType.STRING = parquet.NewStringType()
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("wrong number of children (too few)", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Pass only one child
-		require.False(t, sh.isValidVariantSchema(idx, children[:1]))
-	})
-
-	t.Run("extra children are allowed (for shredding)", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Pass three children (add a fake one that exists in schema list if possible, or just one that is skipped)
-		require.True(t, sh.isValidVariantSchema(idx, append(children, 0)))
-	})
-
-	t.Run("wrong Metadata child name", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Change the Metadata child name
-		sh.Infos[children[0]].InName = "WrongName"
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("wrong Value child name", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Change the Value child name
-		sh.Infos[children[1]].InName = "WrongName"
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("wrong Metadata child type", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Change Metadata type from BYTE_ARRAY to INT32
-		int32Type := parquet.Type_INT32
-		sh.SchemaElements[children[0]].Type = &int32Type
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("wrong Value child type", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Change Value type from BYTE_ARRAY to INT64
-		int64Type := parquet.Type_INT64
-		sh.SchemaElements[children[1]].Type = &int64Type
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("optional Metadata child (should be required)", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Change Metadata repetition type to OPTIONAL
-		optionalRep := parquet.FieldRepetitionType_OPTIONAL
-		sh.SchemaElements[children[0]].RepetitionType = &optionalRep
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("optional Value child (valid for shredded variants)", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		// Change Value repetition type to OPTIONAL - this is valid for shredded variants
-		optionalRep := parquet.FieldRepetitionType_OPTIONAL
-		sh.SchemaElements[children[1]].RepetitionType = &optionalRep
-		require.True(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("nil Metadata type", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		sh.SchemaElements[children[0]].Type = nil
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("nil Value type", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		sh.SchemaElements[children[1]].Type = nil
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("nil Metadata repetition type", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		sh.SchemaElements[children[0]].RepetitionType = nil
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("nil Value repetition type", func(t *testing.T) {
-		sh, idx, children := createVariantSchema()
-		sh.SchemaElements[children[1]].RepetitionType = nil
-		require.False(t, sh.isValidVariantSchema(idx, children))
-	})
-
-	t.Run("child index out of bounds", func(t *testing.T) {
-		sh, idx, _ := createVariantSchema()
-		// Pass children indices that are out of bounds
-		outOfBoundsChildren := []int32{100, 101}
-		require.False(t, sh.isValidVariantSchema(idx, outOfBoundsChildren))
-	})
-}
-
-func TestShreddedVariantSchema(t *testing.T) {
-	// Test that shredded variant schemas (with typed_value) are recognized
-
-	t.Run("valid shredded variant schema with typed_value", func(t *testing.T) {
-		// Create a manual schema with metadata + value (optional) + typed_value
-		requiredRep := parquet.FieldRepetitionType_REQUIRED
-		optionalRep := parquet.FieldRepetitionType_OPTIONAL
-		byteArrayType := parquet.Type_BYTE_ARRAY
-		int32Type := parquet.Type_INT32
-		variantLogicalType := parquet.NewLogicalType()
-		variantLogicalType.VARIANT = parquet.NewVariantType()
-
-		schemas := []*parquet.SchemaElement{
-			{Name: "Root", NumChildren: ptr(int32(1))},
-			{Name: "V", NumChildren: ptr(int32(3)), RepetitionType: &requiredRep, LogicalType: variantLogicalType},
-			{Name: "Metadata", Type: &byteArrayType, RepetitionType: &requiredRep},
-			{Name: "Value", Type: &byteArrayType, RepetitionType: &optionalRep},
-			{Name: "Typed_value", Type: &int32Type, RepetitionType: &optionalRep},
-		}
-
-		sh := NewSchemaHandlerFromSchemaList(schemas)
-		sh.childrenMap = sh.buildChildrenMap()
-
-		// Variant is at index 1
-		idx := int32(1)
-		children := sh.childrenMap[1] // Get children from map
-
-		// Should be valid
-		require.True(t, sh.isValidVariantSchema(idx, children), "Shredded variant schema should be valid")
-
-		// Get schema info
-		info := sh.getVariantSchemaInfo(idx, children)
-		require.NotNil(t, info, "Should get variant schema info")
-		require.True(t, info.IsShredded, "Should be marked as shredded")
-		require.Equal(t, int32(2), info.MetadataIdx, "Metadata index should be 2")
-		require.Equal(t, int32(3), info.ValueIdx, "Value index should be 3")
-		require.Len(t, info.TypedValueIdxs, 1, "Should have one typed_value")
-		require.Equal(t, int32(4), info.TypedValueIdxs[0], "Typed value index should be 4")
-	})
-
-	t.Run("valid shredded variant without value column", func(t *testing.T) {
-		// Create a schema with only metadata + typed_value (no value column)
-		requiredRep := parquet.FieldRepetitionType_REQUIRED
-		optionalRep := parquet.FieldRepetitionType_OPTIONAL
-		byteArrayType := parquet.Type_BYTE_ARRAY
-		int32Type := parquet.Type_INT32
-		variantLogicalType := parquet.NewLogicalType()
-		variantLogicalType.VARIANT = parquet.NewVariantType()
-
-		schemas := []*parquet.SchemaElement{
-			{Name: "Root", NumChildren: ptr(int32(1))},
-			{Name: "V", NumChildren: ptr(int32(2)), RepetitionType: &requiredRep, LogicalType: variantLogicalType},
-			{Name: "Metadata", Type: &byteArrayType, RepetitionType: &requiredRep},
-			{Name: "Typed_value", Type: &int32Type, RepetitionType: &optionalRep},
-		}
-
-		sh := NewSchemaHandlerFromSchemaList(schemas)
-		sh.childrenMap = sh.buildChildrenMap()
-
-		idx := int32(1)
-		children := sh.childrenMap[1]
-
-		require.True(t, sh.isValidVariantSchema(idx, children), "Variant with only metadata+typed_value should be valid")
-
-		info := sh.getVariantSchemaInfo(idx, children)
-		require.NotNil(t, info, "Should get variant schema info")
-		require.True(t, info.IsShredded, "Should be marked as shredded")
-		require.Equal(t, int32(-1), info.ValueIdx, "Value index should be -1 when absent")
-		require.Len(t, info.TypedValueIdxs, 1, "Should have one typed_value")
-	})
-
-	t.Run("typed_value must be OPTIONAL", func(t *testing.T) {
-		// typed_value with REQUIRED repetition should be invalid
-		requiredRep := parquet.FieldRepetitionType_REQUIRED
-		optionalRep := parquet.FieldRepetitionType_OPTIONAL
-		byteArrayType := parquet.Type_BYTE_ARRAY
-		int32Type := parquet.Type_INT32
-		variantLogicalType := parquet.NewLogicalType()
-		variantLogicalType.VARIANT = parquet.NewVariantType()
-
-		schemas := []*parquet.SchemaElement{
-			{Name: "Root", NumChildren: ptr(int32(1))},
-			{Name: "V", NumChildren: ptr(int32(3)), RepetitionType: &requiredRep, LogicalType: variantLogicalType},
-			{Name: "Metadata", Type: &byteArrayType, RepetitionType: &requiredRep},
-			{Name: "Value", Type: &byteArrayType, RepetitionType: &optionalRep},
-			{Name: "Typed_value", Type: &int32Type, RepetitionType: &requiredRep}, // REQUIRED - invalid
-		}
-
-		sh := NewSchemaHandlerFromSchemaList(schemas)
-		sh.childrenMap = sh.buildChildrenMap()
-
-		idx := int32(1)
-		children := sh.childrenMap[1]
-
-		// Still valid because it has value column (typed_value with wrong repetition is just ignored)
-		require.True(t, sh.isValidVariantSchema(idx, children))
-
-		// But typed_value should not be in the info
-		info := sh.getVariantSchemaInfo(idx, children)
-		require.NotNil(t, info)
-		require.Len(t, info.TypedValueIdxs, 0, "REQUIRED typed_value should not be included")
-	})
-}
-
-func TestGetVariantSchemaInfo(t *testing.T) {
-	t.Run("unshredded variant info via index", func(t *testing.T) {
-		sh, err := NewSchemaHandlerFromStruct(new(struct {
-			V any `parquet:"name=variant, type=VARIANT, logicaltype=VARIANT"`
-		}))
-		require.NoError(t, err)
-
-		// Build children map
-		sh.childrenMap = sh.buildChildrenMap()
-
-		// Get info for the variant at index 1
-		idx := int32(1)
-		children := sh.childrenMap[1]
-		info := sh.getVariantSchemaInfo(idx, children)
-
-		require.NotNil(t, info, "Should find variant schema info")
-		require.False(t, info.IsShredded, "Unshredded variant should not be marked as shredded")
-		require.Empty(t, info.TypedValueIdxs, "Should have no typed_value columns")
-		require.True(t, info.ValueIdx >= 0, "Should have a value index")
-	})
-
-	t.Run("invalid schema returns nil", func(t *testing.T) {
-		sh, err := NewSchemaHandlerFromStruct(new(struct {
-			Name string `parquet:"name=name, type=BYTE_ARRAY, convertedtype=UTF8"`
-		}))
-		require.NoError(t, err)
-		sh.childrenMap = sh.buildChildrenMap()
-
-		// Index 1 is Name, not a variant
-		info := sh.getVariantSchemaInfo(1, sh.childrenMap[1])
-		require.Nil(t, info, "Should return nil for non-variant")
-	})
-}
-
-func TestVariantSchemasMap(t *testing.T) {
-	t.Run("VariantSchemas map is populated", func(t *testing.T) {
-		sh, err := NewSchemaHandlerFromStruct(new(struct {
-			V any `parquet:"name=variant, type=VARIANT, logicaltype=VARIANT"`
-		}))
-		require.NoError(t, err)
-
-		require.NotNil(t, sh.VariantSchemas, "VariantSchemas map should be initialized")
-		require.Len(t, sh.VariantSchemas, 1, "Should have one variant")
-
-		// Get the first key in the map
-		var found bool
-		for path := range sh.VariantSchemas {
-			t.Logf("Found variant at path: %s", path)
-			found = true
-		}
-		require.True(t, found, "Variant should be in the map")
-	})
-
-	t.Run("multiple variants are all indexed", func(t *testing.T) {
-		sh, err := NewSchemaHandlerFromStruct(new(struct {
-			V1 any `parquet:"name=variant1, type=VARIANT, logicaltype=VARIANT"`
-			V2 any `parquet:"name=variant2, type=VARIANT, logicaltype=VARIANT"`
-		}))
-		require.NoError(t, err)
-
-		require.Len(t, sh.VariantSchemas, 2, "Should have two variants")
-	})
-}
-
-// Helper function to get pointer to int32
-func ptr(v int32) *int32 {
-	return &v
-}
-
 func TestSchemaHandler_GetVariantSchemaInfo(t *testing.T) {
 	t.Run("returns nil when VariantSchemas is nil", func(t *testing.T) {
 		sh := &SchemaHandler{
@@ -1091,58 +436,6 @@ func TestSchemaHandler_GetVariantSchemaInfo(t *testing.T) {
 			require.Equal(t, expectedInfo, actualInfo)
 		}
 	})
-}
-
-func TestNewSchemaHandlerFromStruct_VariantInfo(t *testing.T) {
-	type VariantType struct {
-		Metadata []byte `parquet:"name=metadata, type=BYTE_ARRAY, encoding=DELTA_LENGTH_BYTE_ARRAY, compression=ZSTD"`
-		Value    []byte `parquet:"name=value, type=BYTE_ARRAY, encoding=DELTA_LENGTH_BYTE_ARRAY, compression=SNAPPY"`
-	}
-
-	type AllTypes struct {
-		Variant VariantType `parquet:"name=Variant, type=VARIANT, logicaltype=VARIANT, logicaltype.specification_version=1"`
-	}
-
-	sh, err := NewSchemaHandlerFromStruct(new(AllTypes))
-	require.NoError(t, err)
-
-	// Parquet_go_root (0)
-	//   Variant (1)
-	//     metadata (2)
-	//     value (3)
-
-	require.Equal(t, int32(2), sh.SchemaElements[1].GetNumChildren())
-	require.NotNil(t, sh.SchemaElements[1].LogicalType)
-	require.NotNil(t, sh.SchemaElements[1].LogicalType.VARIANT)
-
-	// Check metadata field (index 2)
-	require.Equal(t, "Metadata", sh.SchemaElements[2].GetName())
-	require.Equal(t, "metadata", sh.Infos[2].ExName)
-	require.Equal(t, parquet.Encoding_DELTA_LENGTH_BYTE_ARRAY, sh.Infos[2].Encoding)
-	require.NotNil(t, sh.Infos[2].CompressionType)
-	require.Equal(t, parquet.CompressionCodec_ZSTD, *sh.Infos[2].CompressionType)
-
-	// Check value field (index 3)
-	require.Equal(t, "Value", sh.SchemaElements[3].GetName())
-	require.Equal(t, "value", sh.Infos[3].ExName)
-	require.Equal(t, parquet.Encoding_DELTA_LENGTH_BYTE_ARRAY, sh.Infos[3].Encoding)
-	require.NotNil(t, sh.Infos[3].CompressionType)
-	require.Equal(t, parquet.CompressionCodec_SNAPPY, *sh.Infos[3].CompressionType)
-}
-
-func TestNewSchemaHandlerFromStruct_ByteArraySlice(t *testing.T) {
-	type MyStruct struct {
-		Data []byte `parquet:"name=data, type=BYTE_ARRAY"`
-	}
-
-	sh, err := NewSchemaHandlerFromStruct(new(MyStruct))
-	require.NoError(t, err)
-
-	// Root (0)
-	//   Data (1)
-	require.Equal(t, 2, len(sh.SchemaElements))
-	require.Equal(t, int32(0), sh.SchemaElements[1].GetNumChildren())
-	require.Equal(t, parquet.Type_BYTE_ARRAY, *sh.SchemaElements[1].Type)
 }
 
 func TestSchemaHandler_ValidateEncodingsForDataPageVersion(t *testing.T) {
